requirements.txt
ipykernel
jupyter
jupyterlab
psycopg2
pyodbc
pyyaml
pandas
numpy
pathos